+++ CHANGELOG.md
@@ -9,7 +9,11 @@ The internal configuration format may change in minor releases.
 Dependencies are updated to the latest available version during each
 release.  Those changes are not noted here explicitly.
 
-## 5.0.3 (unreleased)
+## 5.1.0 (unreleased)
+
+- Add a Kubernetes `CronJob` to delete entries for expired tokens, note
+  their expiration in the token change history, and truncate history
+  tables.  History entries older than one year are dropped.
 
 - When a user token was edited to change its scope, but not its expiration
   time, its scopes were not updated in Redis.  Since Redis is canonical
@@ -19,13 +23,13 @@ release.  Those changes are not noted here explicitly.
 
 ## 5.0.2 (2022-07-29)
 
-- Improved error handling for LDAP queries.  Hopefully Gafaelfawr should
+- Improve error handling for LDAP queries.  Hopefully Gafaelfawr should
   now recover automatically from LDAP outages.
 
 - Gafaelfawr now uses a pure `pyproject.toml` build system (using the beta
   support in setuptools) rather than using `setup.cfg` or `setup.py`.
 
-- Substantial improvements to the documentation.  Much of the design and
+- Improve the documentation substantially.  Much of the design and
   implementation information has move to tech notes and the Gafaelfawr
   documentation references those.  The changelog is now maintained in
   Markdown to ease preparation of GitHub releases.